feature: dropdown instrument selection, store in config
